move extract_video_metadata to process_geotag_properties.py
